Update GRPCManager connect to support gateway and router
