+++ nym-vpn-apple/ServicesMacOS/Sources/GRPCManager/GRPCManager.swift
@@ -13,16 +13,16 @@ import TunnelStatus
 
 public final class GRPCManager: ObservableObject {
     private let group = MultiThreadedEventLoopGroup(numberOfThreads: 4)
-
     private let channel: GRPCChannel
     private let unixDomainSocket = "/var/run/nym-vpn.sock"
+
     let healthClient: Grpc_Health_V1_HealthClientProtocol
     let client: Nym_Vpn_NymVpndClientProtocol
     let logger = Logger(label: "GRPC Manager")
 
     public static let shared = GRPCManager()
 
-    private var userAgent: Nym_Vpn_UserAgent {
+    var userAgent: Nym_Vpn_UserAgent {
         var agent = Nym_Vpn_UserAgent()
         agent.application = AppVersionProvider.app
         agent.version = "\(AppVersionProvider.appVersion()) (\(daemonVersion))"
@@ -119,70 +119,6 @@ public final class GRPCManager: ObservableObject {
 
             case .failure(let error):
                 self?.logger.log(level: .info, "Failed to connect to VPN: \(error)")
-            }
-        }
-    }
-
-    public func connect(
-        entryGatewayCountryCode: String?,
-        exitRouterCountryCode: String?,
-        isTwoHopEnabled: Bool
-    ) async throws {
-        logger.log(level: .info, "Connecting")
-
-        return try await withCheckedThrowingContinuation { continuation in
-            var request = Nym_Vpn_ConnectRequest()
-            request.userAgent = userAgent
-
-            var entryNode = Nym_Vpn_EntryNode()
-            if let entryGatewayCountryCode {
-                var location = Nym_Vpn_Location()
-                location.twoLetterIsoCountryCode = entryGatewayCountryCode
-                entryNode.location = location
-            } else {
-                // TODO: use it when functionality becomes available
-                //            entryNode.randomLowLatency = Nym_Vpn_Empty()
-                entryNode.random = Nym_Vpn_Empty()
-            }
-
-            var exitNode = Nym_Vpn_ExitNode()
-            if let exitRouterCountryCode {
-                var location = Nym_Vpn_Location()
-                location.twoLetterIsoCountryCode = exitRouterCountryCode
-                exitNode.location = location
-            } else {
-                exitNode.random = Nym_Vpn_Empty()
-            }
-
-            request.entry = entryNode
-            request.exit = exitNode
-
-            request.disableRouting = false
-            request.enableTwoHop = isTwoHopEnabled
-            request.disableBackgroundCoverTraffic = false
-            request.enableCredentialsMode = false
-
-            let call = client.vpnConnect(request, callOptions: nil)
-
-            call.response.whenComplete { [weak self] result in
-                switch result {
-                case .success(let response):
-                    print(response)
-                    self?.logger.log(level: .info, "\(response)")
-
-                    if response.hasError {
-                        if response.error.kind == .noAccountStored {
-                            self?.lastError = GeneralNymError.noMnemonicStored
-                            continuation.resume(throwing: GeneralNymError.noMnemonicStored)
-                        } else {
-                            continuation.resume(throwing: GeneralNymError.library(message: response.error.message))
-                        }
-                    } else {
-                        continuation.resume()
-                    }
-                case .failure(let error):
-                    self?.logger.log(level: .info, "Failed to connect to VPN: \(error)")
-                }
             }
         }
     }
